Word Checker
Created a function that stores variables named for each word and runs if the correct input value is entered. Each variable's value is the set of inputs designated for the word. Each input is converted to a string and then variable value is checked against a string of the correct word